docs(blockvalidation): fix transposed txmeta-store defaults, guard the whole settings table
ProcessTxMetaUsingStoreBatchSize and ProcessTxMetaUsingStoreConcurrency were
documented with each other's values: the batch size is 1024 and the concurrency
is max(4, NumCPU/2), not the other way round. The struct tags were already
correct, so only the markdown table drifted. The high-performance example also
used blockvalidation_processTxMetaUsingStoreBatchSize, which is not a real key
(the key has an underscore before BatchSize), so the example silently did
nothing.

The existing drift guard covered exactly one setting that had never drifted
while two rows in the same table were wrong the whole time. Replace it with a
table-driven check over ExportMetadata() that cross-checks every row's
documented default against the real default, plus a check that every key used
in a configuration example exists. The stale-prose assertions and the
three-way OptimisticMining check are kept - they cover a different failure
mode.
